mmvts/src/requirements.txt
accelerate==0.16.0
cn-clip==1.5.1
datasets==2.9.0
evaluate==0.4.0
fairscale==0.4.13
ffmpeg==1.4
huggingface-hub==4.12.0
matplotlib==3.7.0
multiprocess==0.70.14
nltk==3.8.1
numpy==1.24.2
openai-whisper==20240930
opencv-python==4.9.0.80
packaging==23.0
pandas==1.5.3
pathlib2==2.3.7.post1
Pillow==9.4.0
pip==22.3.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==2.0.0
responses==0.18.0
safetensors==0.4.3
scikit-learn==1.2.1
scipy==1.10.0
segeval==2.0.11
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
setuptools==65.6.3
six==1.16.0
statistics==1.0.3.5
tensorboard==2.14.0
tiktoken==0.7.0
timm==0.9.16
tokenizers==0.12.1
torch==1.12.1+cu102
torchaudio==0.12.1+cu102
torchvision==0.13.1+cu102
transformers==4.26.0